DomainSpecific/dependency/requirements.txt
lxml==5.1.0
#fasttext==0.9.2
fasttext-wheel==0.9.2
sentencepiece==0.1.99
trafilatura==1.6.1
html5lib==1.1
newspaper3k==0.2.8
beautifulsoup4==4.12.2
warcio==1.7.4
markdownify==0.11.6
#cchardet==2.1.7
numpy==1.24.4
scipy==1.10.1
requests==2.32.2
pyarrow==14.0.1
jsonlines==3.1.0
#networkx==3.1
matplotlib==3.7.2
pyyaml==6.0
psutil==5.9.5
tqdm==4.66.3
py_asciimath==0.3.0
pylatexenc==2.10
charset-normalizer==3.2.0
tensorflow==2.12.1
#guesslang==2.2.1
#typing_extensions==4.12.0
faiss-cpu==1.7.4
#torch==2.0.1
#fairscale==0.4.13
sentence_transformers==2.2.2
#PyMuPDF==1.23.6
tiktoken==0.5.2
gensim==4.3.2
openai==1.30.2
boto3==1.34.100
datasets==2.16.0
azure-ai-ml==1.16.0
azure-batch==14.2.0
azure-identity==1.16.1
azure-storage-blob==12.19.1
azure.keyvault.secrets==4.8.0
